+++ packages/libs/lambda-at-edge/src/default-handler.ts
@@ -1354,7 +1354,7 @@ export const generatePermanentPageResponse = async (
     retryStrategy: await buildS3RetryStrategy()
   });
   debug(
-    `[generatePermanentPageResponse] manifest: ${manifest.permanentStaticPages}`
+    `[generatePermanentPageResponse] manifest: ${manifest.permanentStaticPages}.`
   );
   debug(`[generatePermanentPageResponse] uri: ${uri}`);